pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "roof-api"
version = "0.1.0"
description = "API geoespacial para análise de telhados com LIDAR"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [{ name = "RoofApi" }]
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "sqlalchemy>=2.0.0",
    "geoalchemy2>=0.14.0",
    "asyncpg>=0.29.0",
    "alembic>=1.13.0",
    "rasterio>=1.3.0",
    "shapely>=2.0.0",
    "geopandas>=0.14.0",
    "pyproj>=3.6.0",
    "numpy>=1.24.0",
    "scipy>=1.11.0",
    "opencv-python-headless>=4.9.0",
    "scikit-image>=0.22.0",
    "torch>=2.0.0",
    "pillow>=10.0.0",
    "httpx>=0.26.0",
    "python-multipart>=0.0.6",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
]
dml = [
    "torch-directml>=0.2.0",
]
data = [
    "huggingface_hub>=0.20.0",
]

[tool.setuptools.package-dir]
"" = "src"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
